Updated CSS and Generation Code
CSS - Updated the formatting, so elements are now centered and look nice.

Code - Included more data in the main game Object. Moved the increment amount and cost multipliers into it instead of being a hardcoded values.

I still need to come up with a way of adding "juiciness" to these values without creating a wildly different game each time a player progresses in it, or with such a lite experience, maybe this could be an asset?