CSS, Index, and Code
CSS - Referencing new things to bold, like genText and costText.

Index - Separating out the upgrade text from the button(as it was growing to large at higher values), and including in a p tag.

Code - Trying out new calculation values, as well as Flooring them, to ensure that the money value was increasing by hold numbers.